refactor(types): move types into separate files
